Move sapling-specific primitives into the sapling module.
