+++ zcash_proofs/src/prover.rs
@@ -5,11 +5,10 @@ use bls12_381::Bls12;
 use std::path::Path;
 use zcash_primitives::{
     merkle_tree::MerklePath,
-    primitives::{Diversifier, PaymentAddress, ProofGenerationKey, Rseed},
     sapling::{
         prover::TxProver,
         redjubjub::{PublicKey, Signature},
-        Node,
+        Diversifier, Node, PaymentAddress, ProofGenerationKey, Rseed,
     },
     transaction::components::{Amount, GROTH_PROOF_SIZE},
 };
